requirements.txt
pytest==9.0.3
requests==2.33.1
python-dotenv==1.2.2
pytest-html==4.2.0